moved things places
